wip: refinements
