+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/hyperboliclabs/ai/__init__.py
@@ -6,9 +6,11 @@ including text, image, and audio generation.
 
 from .action_provider import HyperbolicAIActionProvider, hyperbolic_ai_action_provider
 from .service import AIService
+from .utils import save_base64_image
 
 __all__ = [
     "HyperbolicAIActionProvider",
     "hyperbolic_ai_action_provider",
     "AIService",
+    "save_base64_image",
 ]
