+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/hyperboliclabs/ai/service.py
@@ -9,7 +9,7 @@ from .models import (
     ImageGenerationRequest,
     ImageGenerationResponse,
 )
-from ..services.base import Base
+from ..service import Base
 
 
 class AIService(Base):
