+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/hyperboliclabs/settings/__init__.py
@@ -1,0 +1,17 @@
+"""Hyperbolic settings services.
+
+This package provides modules for interacting with Hyperbolic settings services.
+It includes models, schemas, and utility functions for settings operations like wallet linking.
+"""
+
+from .action_provider import HyperbolicSettingsActionProvider
+from .models import WalletLinkRequest, WalletLinkResponse
+from .utils import format_wallet_link_response
+
+
+__all__ = [
+    "HyperbolicSettingsActionProvider",
+    "WalletLinkRequest",
+    "WalletLinkResponse",
+    "format_wallet_link_response",
+]
