+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/hyperboliclabs/settings/service.py
@@ -3,7 +3,8 @@
 from typing import Any
 
 from ..constants import SETTINGS_BASE_URL, SETTINGS_ENDPOINTS
-from .base import Base
+from ..service import Base
+from .models import WalletLinkResponse
 
 
 class Settings(Base):
@@ -18,19 +19,21 @@ class Settings(Base):
         """
         super().__init__(api_key, SETTINGS_BASE_URL)
 
-    def link_wallet(self, wallet_address: str) -> dict[str, Any]:
+    def link_wallet(self, wallet_address: str) -> WalletLinkResponse:
         """Link a wallet address to the Hyperbolic account.
 
         Args:
             wallet_address: The wallet address to link.
 
         Returns:
-            dict[str, Any]: The wallet linking response data.
+            WalletLinkResponse: The wallet linking response data.
 
         Raises:
             requests.exceptions.RequestException: If the API request fails.
 
         """
-        return self.make_request(
+        response_data = self.make_request(
             endpoint=SETTINGS_ENDPOINTS["LINK_WALLET"], data={"wallet_address": wallet_address}
-        ) 
+        )
+        
+        return WalletLinkResponse.model_validate(response_data)
