+++ python/coinbase-agentkit/tests/action_providers/hyperboliclabs/ai/conftest.py
@@ -1,24 +1,11 @@
 """Test fixtures for Hyperbolic AI service."""
 
-import os
 from unittest.mock import patch
 
 import pytest
 
 from coinbase_agentkit.action_providers.hyperboliclabs.ai.service import AIService
-
-# Test constants
-TEST_API_KEY = ""
-
-
-@pytest.fixture
-def api_key() -> str:
-    """Get API key for testing.
-    
-    Returns:
-        str: API key from environment or default test key.
-    """
-    return os.environ.get("HYPERBOLIC_API_KEY", TEST_API_KEY)
+from coinbase_agentkit.action_providers.hyperboliclabs.ai.action_provider import HyperbolicAIActionProvider
 
 
 @pytest.fixture
@@ -35,9 +22,35 @@ def ai_service(api_key: str) -> AIService:
 
 
 @pytest.fixture
+def mock_ai_service():
+    """Create a mock AIService for testing.
+    
+    Returns:
+        MagicMock: A mock object that simulates the AIService.
+    """
+    with patch("coinbase_agentkit.action_providers.hyperboliclabs.ai.action_provider.AIService") as mock:
+        yield mock.return_value
+
+
+@pytest.fixture
+def provider(mock_ai_service):
+    """Create a HyperbolicAIActionProvider with a test API key and mock service.
+    
+    Args:
+        mock_ai_service: Mock AIService to use in the provider.
+        
+    Returns:
+        HyperbolicAIActionProvider: Provider with mock service.
+    """
+    provider = HyperbolicAIActionProvider(api_key="test-api-key")
+    provider.ai_service = mock_ai_service
+    return provider
+
+
+@pytest.fixture
 def mock_request():
     """Mock requests for all tests."""
-    with patch("requests.request") as mock:
+    with patch("coinbase_agentkit.action_providers.hyperboliclabs.service.requests.request") as mock:
         mock.return_value.json.return_value = {"status": "success"}
         mock.return_value.raise_for_status.return_value = None
         yield mock
