+++ python/coinbase-agentkit/tests/action_providers/hyperboliclabs/billing/test_service.py
@@ -6,18 +6,16 @@ import requests
 from coinbase_agentkit.action_providers.hyperboliclabs.constants import BILLING_BASE_URL
 from coinbase_agentkit.action_providers.hyperboliclabs.billing.service import Billing
 
-from .conftest import TEST_API_KEY
 
-
-def test_billing_service_init():
+def test_billing_service_init(api_key):
     """Test Billing service initialization."""
-    service = Billing(TEST_API_KEY)
+    service = Billing(api_key)
     assert service.base_url == BILLING_BASE_URL
 
 
-def test_billing_get_current_balance(mock_request):
+def test_billing_get_current_balance(mock_request, api_key):
     """Test get_balance method."""
-    service = Billing(TEST_API_KEY)
+    service = Billing(api_key)
     mock_request.return_value.json.return_value = {"credits": "1000.50"}
 
     response = service.get_balance()
@@ -25,9 +23,9 @@ def test_billing_get_current_balance(mock_request):
     mock_request.assert_called_once()
 
 
-def test_billing_get_purchase_history(mock_request):
+def test_billing_get_purchase_history(mock_request, api_key):
     """Test get_purchase_history method."""
-    service = Billing(TEST_API_KEY)
+    service = Billing(api_key)
     mock_request.return_value.json.return_value = {
         "purchase_history": [
             {
@@ -46,9 +44,9 @@ def test_billing_get_purchase_history(mock_request):
     mock_request.assert_called_once()
 
 
-def test_billing_service_error_handling(mock_request):
+def test_billing_service_error_handling(mock_request, api_key):
     """Test error handling in billing service."""
-    service = Billing(TEST_API_KEY)
+    service = Billing(api_key)
     mock_request.side_effect = requests.exceptions.HTTPError("403 Forbidden: Insufficient credits")
 
     with pytest.raises(requests.exceptions.HTTPError, match="403 Forbidden"):
