+++ python/coinbase-agentkit/tests/action_providers/hyperboliclabs/marketplace/test_service.py
@@ -22,7 +22,6 @@ from coinbase_agentkit.action_providers.hyperboliclabs.marketplace.models import
 from coinbase_agentkit.action_providers.hyperboliclabs.marketplace.service import Marketplace
 
 from .conftest import (
-    TEST_API_KEY,
     TEST_CLUSTER,
     TEST_GPU_COUNT,
     TEST_INSTANCE_ID,
@@ -30,16 +29,16 @@ from .conftest import (
 )
 
 
-def test_marketplace_init():
+def test_marketplace_init(api_key):
     """Test Marketplace service initialization."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     assert service.base_url == MARKETPLACE_BASE_URL
-    assert service.api_key == TEST_API_KEY
+    assert service.api_key == api_key
 
 
-def test_marketplace_get_available_instances(mock_request):
+def test_marketplace_get_available_instances(mock_request, api_key):
     """Test get_available_instances method with empty response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.return_value.json.return_value = {"instances": []}
 
     response = service.get_available_instances()
@@ -53,9 +52,9 @@ def test_marketplace_get_available_instances(mock_request):
     assert mock_request.call_args.kwargs.get("json") == {"filters": {}}
 
 
-def test_marketplace_get_available_instances_with_data(mock_request):
+def test_marketplace_get_available_instances_with_data(mock_request, api_key):
     """Test get_available_instances method with populated response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "instances": [
@@ -86,9 +85,9 @@ def test_marketplace_get_available_instances_with_data(mock_request):
     mock_request.assert_called_once()
 
 
-def test_marketplace_get_instance_history(mock_request):
+def test_marketplace_get_instance_history(mock_request, api_key):
     """Test get_instance_history method with empty response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.return_value.json.return_value = {"instance_history": []}
 
     response = service.get_instance_history()
@@ -101,9 +100,9 @@ def test_marketplace_get_instance_history(mock_request):
     assert "Bearer" in mock_request.call_args.kwargs.get("headers", {}).get("Authorization", "")
 
 
-def test_marketplace_get_instance_history_with_data(mock_request):
+def test_marketplace_get_instance_history_with_data(mock_request, api_key):
     """Test get_instance_history method with populated response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "instance_history": [
@@ -133,9 +132,9 @@ def test_marketplace_get_instance_history_with_data(mock_request):
     mock_request.assert_called_once()
 
 
-def test_marketplace_get_rented_instances(mock_request):
+def test_marketplace_get_rented_instances(mock_request, api_key):
     """Test get_rented_instances method with empty response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.return_value.json.return_value = {"instances": []}
 
     response = service.get_rented_instances()
@@ -148,9 +147,9 @@ def test_marketplace_get_rented_instances(mock_request):
     assert "Bearer" in mock_request.call_args.kwargs.get("headers", {}).get("Authorization", "")
 
 
-def test_marketplace_get_rented_instances_with_data(mock_request):
+def test_marketplace_get_rented_instances_with_data(mock_request, api_key):
     """Test get_rented_instances method with populated response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "instances": [
@@ -186,9 +185,9 @@ def test_marketplace_get_rented_instances_with_data(mock_request):
     mock_request.assert_called_once()
 
 
-def test_marketplace_rent_instance_success(mock_request):
+def test_marketplace_rent_instance_success(mock_request, api_key):
     """Test rent_instance method with successful response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "status": "success",
@@ -219,9 +218,9 @@ def test_marketplace_rent_instance_success(mock_request):
     assert "image" in json_data
 
 
-def test_marketplace_rent_instance_with_message(mock_request):
+def test_marketplace_rent_instance_with_message(mock_request, api_key):
     """Test rent_instance method with additional message in response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "status": "success",
@@ -252,9 +251,9 @@ def test_marketplace_rent_instance_with_message(mock_request):
     assert "image" in json_data
 
 
-def test_marketplace_rent_instance_error(mock_request):
+def test_marketplace_rent_instance_error(mock_request, api_key):
     """Test rent_instance method with HTTP error response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.side_effect = requests.exceptions.HTTPError("400 Client Error: Bad Request")
 
     request = RentInstanceRequest(
@@ -266,9 +265,9 @@ def test_marketplace_rent_instance_error(mock_request):
         service.rent_instance(request)
 
 
-def test_marketplace_rent_instance_connection_error(mock_request):
+def test_marketplace_rent_instance_connection_error(mock_request, api_key):
     """Test rent_instance method with connection error."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.side_effect = requests.exceptions.ConnectionError("Connection refused")
     
     request = RentInstanceRequest(
@@ -280,9 +279,9 @@ def test_marketplace_rent_instance_connection_error(mock_request):
         service.rent_instance(request)
 
 
-def test_marketplace_terminate_instance_success(mock_request):
+def test_marketplace_terminate_instance_success(mock_request, api_key):
     """Test terminate_instance method with successful response."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "status": "success",
@@ -302,9 +301,9 @@ def test_marketplace_terminate_instance_success(mock_request):
     assert mock_request.call_args.kwargs.get("json") == {"id": TEST_INSTANCE_ID}
 
 
-def test_marketplace_terminate_instance_with_custom_message(mock_request):
+def test_marketplace_terminate_instance_with_custom_message(mock_request, api_key):
     """Test terminate_instance method with custom success message."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     
     mock_request.return_value.json.return_value = {
         "status": "success",
@@ -323,9 +322,9 @@ def test_marketplace_terminate_instance_with_custom_message(mock_request):
     assert mock_request.call_args.kwargs.get("json") == {"id": "i-abc123"}
 
 
-def test_marketplace_terminate_instance_not_found(mock_request):
+def test_marketplace_terminate_instance_not_found(mock_request, api_key):
     """Test terminate_instance method with not found error."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.side_effect = requests.exceptions.HTTPError("404 Client Error: Not Found")
 
     request = TerminateInstanceRequest(id="invalid-id")
@@ -333,9 +332,9 @@ def test_marketplace_terminate_instance_not_found(mock_request):
         service.terminate_instance(request)
 
 
-def test_marketplace_terminate_instance_bad_request(mock_request):
+def test_marketplace_terminate_instance_bad_request(mock_request, api_key):
     """Test terminate_instance method with bad request error."""
-    service = Marketplace(TEST_API_KEY)
+    service = Marketplace(api_key)
     mock_request.side_effect = requests.exceptions.HTTPError("400 Client Error: Bad Request")
     
     request = TerminateInstanceRequest(id="")
