+++ python/coinbase-agentkit/tests/action_providers/hyperboliclabs/marketplace/test_service_e2e_fixtures.py
@@ -5,7 +5,7 @@ import time
 
 import pytest
 
-from coinbase_agentkit.action_providers.hyperboliclabs.models import (
+from coinbase_agentkit.action_providers.hyperboliclabs.marketplace.models import (
     RentInstanceRequest,
     TerminateInstanceRequest,
 )
