--- python/coinbase-agentkit/tests/action_providers/hyperboliclabs/services/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for Hyperbolic services."""
